+++ packages/solid-query/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @tanstack/solid-query
 
+## 5.98.0
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @tanstack/query-core@5.98.0
+
+## 5.97.0
+
+### Patch Changes
+
+- Updated dependencies [[`2bfb12c`](https://github.com/TanStack/query/commit/2bfb12cc44f1d8495106136e4ddacb817135f8f9)]:
+  - @tanstack/query-core@5.97.0
+
 ## 5.96.2
 
 ### Patch Changes