+++ packages/svelte-query-devtools/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @tanstack/svelte-query-devtools
 
+## 6.1.15
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @tanstack/query-devtools@5.98.0
+  - @tanstack/svelte-query@6.1.15
+
+## 6.1.14
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @tanstack/query-devtools@5.97.0
+  - @tanstack/svelte-query@6.1.14
+
 ## 6.1.13
 
 ### Patch Changes