+++ packages/vue-query/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @tanstack/vue-query
 
+## 5.98.0
+
+### Minor Changes
+
+- Add usePrefetchQuery and usePrefetchInfiniteQuery to vue-query. ([#10372](https://github.com/TanStack/query/pull/10372))
+
+### Patch Changes
+
+- fix(vue-query): fix type of queryOptions to allow plain properies or getters ([#10452](https://github.com/TanStack/query/pull/10452))
+
+- Updated dependencies []:
+  - @tanstack/query-core@5.98.0
+
+## 5.97.0
+
+### Patch Changes
+
+- Updated dependencies [[`2bfb12c`](https://github.com/TanStack/query/commit/2bfb12cc44f1d8495106136e4ddacb817135f8f9)]:
+  - @tanstack/query-core@5.97.0
+
 ## 5.96.2
 
 ### Patch Changes